+++ src/main/java/kr/codesqaud/cafe/controller/ArticleController.java
@@ -1,6 +1,6 @@
 package kr.codesqaud.cafe.controller;
 
-import kr.codesqaud.cafe.dto.ArticleDto;
+import kr.codesqaud.cafe.controller.dto.ArticleDto;
 import kr.codesqaud.cafe.service.ArticleService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -24,7 +24,7 @@ public class ArticleController {
     }
 
     @GetMapping("/articles/{id}")
-    public String showUserProfile(@PathVariable int id, Model model) {
+    public String showDetailArticle(@PathVariable int id, Model model) {
         model.addAttribute("article", articleService.findArticleById(id));
         return "/post/show";
     }
